requirements.txt
flask
requests
discord.py
Gunicorn